+++ relay-server/src/utils/mod.rs
@@ -25,6 +25,7 @@ mod memory;
 #[cfg(feature = "processing")]
 mod native;
 mod serde;
+#[cfg(feature = "processing")]
 mod unreal;
 
 pub use self::api::*;
@@ -48,4 +49,5 @@ pub use self::split_off::*;
 pub use self::statsd::*;
 pub use self::stream::*;
 pub use self::thread_pool::*;
+#[cfg(feature = "processing")]
 pub use self::unreal::*;