+++ relay-server/src/utils/unreal.rs
@@ -55,8 +55,10 @@ pub fn extract_items(payload: Bytes, config: &Config) -> Result<Items, Processin
     Ok(items)
 }
 
-/// Expands items previously extracted from a report in to the [`UnrealExpansion`] representation.
-pub fn expand_unreal_items(items: Items) -> Result<UnrealExpansion, ProcessingError> {
+/// Expands an Unreal 4 crash report payload and returns the expanded items.
+pub fn expand_unreal(payload: Bytes, config: &Config) -> Result<UnrealExpansion, ProcessingError> {
+    let items = extract_items(payload, config)?;
+
     let mut context = items
         .iter()
         .find(|&item| matches!(item.attachment_type(), Some(AttachmentType::UnrealContext)))
@@ -72,25 +74,15 @@ pub fn expand_unreal_items(items: Items) -> Result<UnrealExpansion, ProcessingEr
     })
 }
 
-/// Expands an Unreal 4 crash report payload and returns the expanded items.
-#[cfg_attr(not(feature = "processing"), expect(unused))]
-pub fn expand_unreal(payload: Bytes, config: &Config) -> Result<UnrealExpansion, ProcessingError> {
-    let attachments = extract_items(payload, config)?;
-    expand_unreal_items(attachments)
-}
-
 /// Expansion from an Unreal 4 report.
 pub struct UnrealExpansion {
     /// The error event if the crash contained one.
-    #[cfg_attr(not(feature = "processing"), expect(unused))]
     pub event: Option<Item>,
     /// The parsed unreal context.
     ///
     /// Note: the raw unreal context may still be in [`Self::attachments`].
-    #[cfg_attr(not(feature = "processing"), expect(unused))]
     pub context: Option<Unreal4Context>,
     /// Files of the report as attachments.
-    #[cfg_attr(not(feature = "processing"), expect(unused))]
     pub attachments: Items,
 }
 
@@ -352,7 +344,6 @@ fn merge_unreal_context(event: &mut Event, context: Unreal4Context) {
 /// Processes an unreal crash report.
 ///
 /// The `user_header` should be extracted from the [`crate::constants::UNREAL_USER_HEADER`] envelope header.
-#[cfg_attr(not(feature = "processing"), expect(unused))]
 pub fn process_unreal<'a>(
     context: Option<Unreal4Context>,
     event_id: EventId,
@@ -394,7 +385,6 @@ pub fn process_unreal<'a>(
 }
 
 /// Result when processing an unreal report.
-#[cfg_attr(not(feature = "processing"), expect(unused))]
 pub struct ProcessedUnrealReport {
     /// User reports contained in the report.
     pub user_reports: Items,